upd: refactoring
