+++ .gitignore
@@ -7,4 +7,5 @@ bin
 .idea
 CMakeFiles
 CMakeUser*
+
 solution
