+++ src/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_subdirectory(lib)
+
+add_executable(test_app test.cpp)
+target_link_libraries(test_app order order_analytics)
