+++ src/test.cpp
@@ -1,0 +1,8 @@
+#include <order.hpp>
+#include <order_analytics.hpp>
+
+int main() {
+    OrderAnalytics orders("format.json");
+    orders.CreateMonthReport();
+    return 0;
+}
